Consolidated files and folders into the app folder.
