+++ app/ui/components/input.tsx
@@ -1,5 +1,5 @@
 import * as React from "react"
-import { cn } from "@/lib/utils"
+import { cn } from "@/app/lib/utils"
 
 export type InputProps = React.InputHTMLAttributes<HTMLInputElement>
