+++ app/ui/components/slider.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 import * as SliderPrimitive from "@radix-ui/react-slider"
-import { cn } from "@/lib/utils"
+import { cn } from "@/app/lib/utils"
 
 const CustomSlider = React.forwardRef<
   React.ElementRef<typeof SliderPrimitive.Root>,
