+++ app/ui/theme-toggle.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import { useEffect, useState } from "react";
 import { FiSun, FiMoon, FiMonitor } from "react-icons/fi";
